+++ src/components/multi-select/MultiSelect.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import clsx from "clsx";
 import { FocusScope } from "react-aria";
 import {
     Button,
@@ -13,15 +12,18 @@ import {
     TextField,
 } from "react-aria-components";
 import { ErrorBoundary } from "react-error-boundary";
-import { twMerge } from "tailwind-merge";
 
 import { XCircleIcon } from "@heroicons/react/16/solid";
 import { ArrowPathIcon } from "@heroicons/react/24/solid";
 
 import { MultiSelectProps } from ".";
 import useMultiSelect from "./useMultiSelect";
+import useMultiSelectClasses from "./useMultiSelectClasses";
 
 const MultiSelect = <T extends object>(props: MultiSelectProps<T>) => {
+    /**
+     * Destructure props
+     */
     const {
         autoFocus,
         classes,
@@ -34,6 +36,9 @@ const MultiSelect = <T extends object>(props: MultiSelectProps<T>) => {
         selectedItems,
         ...rest
     } = props;
+    /**
+     * Destructure hooks
+     */
     const {
         errorMessage,
         inputFieldRef,
@@ -59,48 +64,23 @@ const MultiSelect = <T extends object>(props: MultiSelectProps<T>) => {
         openList,
         removeItem,
     } = useMultiSelect(props);
-    const labelCN = twMerge(
-        clsx(
-            "relative m-0 p-0 px-0.5 text-xs leading-none text-slate-400 [&>.label-line]:bg-white",
-            classes?.label
-        )
-    );
-    const labelWrapperCN = twMerge(
-        clsx(
-            "absolute left-0 top-0 z-[2] flex transition-opacity duration-150 ease-in-out",
-            classes?.labelWrapper,
-            {
-                "pointer-events-none opacity-0": !filterText,
-                "pointer-events-auto opacity-100":
-                    filterText || placeholder || selectedItems?.length > 0,
-            }
-        )
-    );
-    const listCN = twMerge(
-        clsx(
-            "max-h-[500px] min-h-[150px] h-[calc(100vh/2-100px)] overflow-auto bg-white max-w-full",
-            classes?.list
-        )
-    );
-    const listItemCN = twMerge(
-        clsx(
-            "selected:bg-slate-100 relative cursor-default px-2 py-1 outline-none focus-visible:bg-slate-200",
-            classes?.listItem
-        )
-    );
-    const popoverCN = twMerge(clsx("max-w-full", classes?.popover));
-    const selectCN = twMerge(
-        clsx(
-            "relative max-h-72 w-full overflow-auto rounded-md border border-slate-200 bg-white/90 px-1.5 py-2 shadow",
-            classes?.select
-        )
-    );
-    const tagCN = twMerge(
-        clsx(
-            "item-center inline-flex gap-1 rounded-md border border-slate-200 bg-slate-50 px-2 py-1 text-sm text-slate-500 selected:border-blue-400",
-            classes?.tag
-        )
-    );
+    /**
+     * Destructure classes
+     */
+    const {
+        labelCN,
+        labelWrapperCN,
+        listCN,
+        listItemCN,
+        popoverCN,
+        selectCN,
+        tagCN,
+    } = useMultiSelectClasses({
+        classes,
+        filterText,
+        selectedItems,
+        placeholder,
+    });
 
     return (
         <ErrorBoundary fallback={<p>⚠️Something went wrong</p>}>
